Mutex Works
Replay Mutex state via Sync Op record and replay.
Using the naive solution for now because when CondVars
get replayed they will need locks to be held, so handling
that optimization will require some actual thought to get it done properly.

Uses `cur_count` in exec_env as an ID.

+++ include/wamr_exec_env.h
@@ -18,7 +18,7 @@ struct WAMRExecEnv { // multiple
     //
     //    /* Previous thread's exec env of a WASM module instance. */
     //    struct WASMExecEnv *prev;
-    uint8 cur_count{};
+    ssize_t cur_count{};
 
     /* Note: field module_inst, argv_buf, native_stack_boundary,
        susƒend_flags, aux_stack_boundary, aux_stack_bottom, and
